Contrastes post hoc de una matriz de datos con chi-cuadrada

Muestra las comparaciones post-hoc por pares de una prueba Chi cuadrada significativa. Calcula los contrastes si hay más de dos grupos. Se aplica una prueba de fisher individual para saber que grupos difieren. Los valores p ajustados pueden calcularse con una amplia variedad de métodos: fdr, BH, BY, bonferroni, holm, hochberg y hommel. Funciona básicamente como una función envolvente que envía los valores p "brutos" no ajustados de cada prueba de chi-cuadrado por pares a la función p.adjust en el programa R base. La función p.adjust debe ser consultada para una mayor descripción de los métodos utilizados. Función tomada y modificada de la paqueteria fifer. https://www.rdocumentation.org/packages/fifer/versions/1.1

Usage

chisq.post.hoc(
  tabla,
  test = c("fisher.test"),
  popsInRows = TRUE,
  control = c("fdr", "BH", "BY", "bonferroni", "holm", "hochberg", "hommel"),
  digits = 4,
  simulate.p.value = NULL
)

Arguments

tabla

Una tabla o matriz usada para la función chisq.test

test

Método usado para ajustar el valor de P.

popsInRows

Un argumento lógico que indica si las poblaciones forman las filas de la tabla (por defecto =TRUE) o no (=FALSE).

control

Indica el método de control a utilizar. fdr, BH, BY, bonferroni, holm, hochberg y hommel.

digits

Número de digitos a mostrar.

simulate.p.value

Si argumento es TRUE, indica que se calculan los valores p mediante la simulación de Monte Carlo y 5000 replicas. Para más información consulta la funcion 'chisq.test'.

Value

Contrastes 'post-hoc' de chi cuadrada para dos niveles usando Fisher.test.

Examples

table <- as.table(rbind(c(14, 43), c(23, 47)))
rownames(table) <- c("Hembras", "Machos")
chisq.post.hoc(table,test = c("fisher.test"))
#> 
#> Los valores de p fueron ajustados con el metodo: fdr
#> 
#> 
#> Table: --
#> 
#> | Comparar          |     p| p.ajust|
#> |------------------:|-----:|-------:|
#> | Hembras vs. Machos| 0.332|   0.332|

xtab <- as.table(rbind(c(180, 145), c(179, 106),c(510, 196), c(862, 23)))
dimnames(xtab) <- list(
Tratamiento = c("1st", "2nd", "3rd", "Control"),
Genero = c("hembra", "macho"))
chisq.post.hoc(xtab)
#> 
#> Los valores de p fueron ajustados con el metodo: fdr
#> 
#> 
#> Table: --
#> 
#> | Comparar       |      p| p.ajust|
#> |---------------:|------:|-------:|
#> | 1st vs. 2nd    |  0.069|   0.069|
#> | 1st vs. 3rd    | <0.001|  <0.001|
#> | 1st vs. Control| <0.001|  <0.001|
#> | 2nd vs. 3rd    |  0.004|   0.005|
#> | 2nd vs. Control| <0.001|  <0.001|
#> | 3rd vs. Control| <0.001|  <0.001|
chisq.post.hoc(xtab,control = c("fdr"))
#> 
#> Los valores de p fueron ajustados con el metodo: fdr
#> 
#> 
#> Table: --
#> 
#> | Comparar       |      p| p.ajust|
#> |---------------:|------:|-------:|
#> | 1st vs. 2nd    |  0.069|   0.069|
#> | 1st vs. 3rd    | <0.001|  <0.001|
#> | 1st vs. Control| <0.001|  <0.001|
#> | 2nd vs. 3rd    |  0.004|   0.005|
#> | 2nd vs. Control| <0.001|  <0.001|
#> | 3rd vs. Control| <0.001|  <0.001|
chisq.post.hoc(xtab,control = c("BY"))
#> 
#> Los valores de p fueron ajustados con el metodo: BY
#> 
#> 
#> Table: --
#> 
#> | Comparar       |      p| p.ajust|
#> |---------------:|------:|-------:|
#> | 1st vs. 2nd    |  0.069|   0.171|
#> | 1st vs. 3rd    | <0.001|  <0.001|
#> | 1st vs. Control| <0.001|  <0.001|
#> | 2nd vs. 3rd    |  0.004|   0.013|
#> | 2nd vs. Control| <0.001|  <0.001|
#> | 3rd vs. Control| <0.001|  <0.001|
chisq.post.hoc(xtab,control = c("bonferroni"))
#> 
#> Los valores de p fueron ajustados con el metodo: bonferroni
#> 
#> 
#> Table: --
#> 
#> | Comparar       |      p| p.ajust|
#> |---------------:|------:|-------:|
#> | 1st vs. 2nd    |  0.069|   0.418|
#> | 1st vs. 3rd    | <0.001|  <0.001|
#> | 1st vs. Control| <0.001|  <0.001|
#> | 2nd vs. 3rd    |  0.004|   0.028|
#> | 2nd vs. Control| <0.001|  <0.001|
#> | 3rd vs. Control| <0.001|  <0.001|
chisq.post.hoc(xtab,control = c("bonferroni"), simulate.p.value = TRUE)
#> Los valores de p fueron ajustados con el metodobonferroni
#> 
#> Los valores de p fueron ajustados con el metodo: bonferroni usando 5000 simulaciones.
#> 
#> 
#> Table: --
#> 
#> | Comparar       |      p| p.ajust|
#> |---------------:|------:|-------:|
#> | 1st vs. 2nd    |  0.069|   0.418|
#> | 1st vs. 3rd    | <0.001|  <0.001|
#> | 1st vs. Control| <0.001|  <0.001|
#> | 2nd vs. 3rd    |  0.004|   0.028|
#> | 2nd vs. Control| <0.001|  <0.001|
#> | 3rd vs. Control| <0.001|  <0.001|