Contrastes post hoc de una matriz de datos con chi-cuadrada

Muestra las comparaciones post-hoc por pares de una prueba Chi cuadrada. Realiza la prueba exacta de Fisher para comprobar independencia de filas en una tabla de contingencia, luego de obtener una prueba chi-cuadrada significativa.

Usage

chisq.post.hoc.filas(tabla, p.adjust.method = NULL, alternative = NULL)

Arguments

tabla

Una tabla o matriz, la misma que se usó con la función chisq.test

p.adjust.method

Método usado para ajustar el valor de P. Se puede usar: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", y "none".

alternative

Un argumento que indica la hipótesis alternativa que se probará.

Value

Contrastes 'post-hoc' de chi cuadrada para dos o mas filas usando la prueba Fisher.test.

Details

Función tomada y modificada de la paqueteria rstatix. https://cran.r-project.org/web/packages/rstatix/index.html

Examples

xtab <- as.table(rbind(c(180, 145), c(179, 106),c(510, 196), c(862, 23)))
dimnames(xtab) <- list(
Tratamiento = c("1st", "2nd", "3rd", "Control"),
Genero = c("hembra", "macho"))
chisq.post.hoc.filas(xtab)
#> 
#> Los valores de p fueron ajustados con el metodo: bonferroni, y la hipótesis alternativa: two.sided
#> 
#> 
#> Table: --
#> 
#> | grupo  |    n|      p| p.ajust.|
#> |-------:|----:|------:|--------:|
#> | 1st    | 2201| <0.001|   <0.001|
#> | 2nd    | 2201| <0.001|   <0.001|
#> | 3rd    | 2201| <0.001|   <0.001|
#> | Control| 2201| <0.001|   <0.001|
#solo hipótesis alternativa
chisq.post.hoc.filas(xtab, alternative = "less")
#> 
#> Los valores de p fueron ajustados con el metodo: bonferroni y la hipótesis alternativa: less
#> 
#> 
#> Table: --
#> 
#> | grupo  |    n|      p| p.ajust.|
#> |-------:|----:|------:|--------:|
#> | 1st    | 2201| <0.001|   <0.001|
#> | 2nd    | 2201| <0.001|   <0.001|
#> | 3rd    | 2201| <0.001|   <0.001|
#> | Control| 2201|      1|        1|
chisq.post.hoc.filas(xtab, alternative = "greater")
#> 
#> Los valores de p fueron ajustados con el metodo: bonferroni y la hipótesis alternativa: greater
#> 
#> 
#> Table: --
#> 
#> | grupo  |    n|      p| p.ajust.|
#> |-------:|----:|------:|--------:|
#> | 1st    | 2201|      1|        1|
#> | 2nd    | 2201|      1|        1|
#> | 3rd    | 2201|      1|        1|
#> | Control| 2201| <0.001|   <0.001|
chisq.post.hoc.filas(xtab, alternative = "two.sided")
#> 
#> Los valores de p fueron ajustados con el metodo: bonferroni y la hipótesis alternativa: two.sided
#> 
#> 
#> Table: --
#> 
#> | grupo  |    n|      p| p.ajust.|
#> |-------:|----:|------:|--------:|
#> | 1st    | 2201| <0.001|   <0.001|
#> | 2nd    | 2201| <0.001|   <0.001|
#> | 3rd    | 2201| <0.001|   <0.001|
#> | Control| 2201| <0.001|   <0.001|
#con los dos argumentos
chisq.post.hoc.filas(xtab, p.adjust.method = "BY", alternative = "less")
#> 
#> Los valores de p fueron ajustados con el metodo: BY y la hipótesis alternativa: less
#> 
#> 
#> Table: --
#> 
#> | grupo  |    n|      p| p.ajust.|
#> |-------:|----:|------:|--------:|
#> | 1st    | 2201| <0.001|   <0.001|
#> | 2nd    | 2201| <0.001|   <0.001|
#> | 3rd    | 2201| <0.001|   <0.001|
#> | Control| 2201|      1|        1|
chisq.post.hoc.filas(xtab, p.adjust.method = "BY", alternative = "two.sided")
#> 
#> Los valores de p fueron ajustados con el metodo: BY y la hipótesis alternativa: two.sided
#> 
#> 
#> Table: --
#> 
#> | grupo  |    n|      p| p.ajust.|
#> |-------:|----:|------:|--------:|
#> | 1st    | 2201| <0.001|   <0.001|
#> | 2nd    | 2201| <0.001|   <0.001|
#> | 3rd    | 2201| <0.001|   <0.001|
#> | Control| 2201| <0.001|   <0.001|
chisq.post.hoc.filas(xtab, p.adjust.method = "holm", alternative = "less")
#> 
#> Los valores de p fueron ajustados con el metodo: holm y la hipótesis alternativa: less
#> 
#> 
#> Table: --
#> 
#> | grupo  |    n|      p| p.ajust.|
#> |-------:|----:|------:|--------:|
#> | 1st    | 2201| <0.001|   <0.001|
#> | 2nd    | 2201| <0.001|   <0.001|
#> | 3rd    | 2201| <0.001|   <0.001|
#> | Control| 2201|      1|        1|