Formato de valores de p

Functions to fromat p values.

Usage

format_p_values(
  p,
  digits_p = 3,
  cols = NULL,
  ...,
  alpha = 0.05,
  signif_stars = TRUE,
  rm_zero = FALSE,
  add_p = FALSE,
  rm_spaces = FALSE,
  ss = signif_syms
)

# S3 method for default
format_p_values(
  p,
  digits_p = 3,
  cols = NULL,
  ...,
  signif_stars = TRUE,
  rm_zero = FALSE,
  add_p = FALSE,
  rm_spaces = FALSE,
  ss = signif_syms
)

format_p(
  p_i,
  digits_p = 3,
  signif_stars = TRUE,
  rm_zero = FALSE,
  add_p = FALSE,
  rm_spaces = FALSE,
  ss = signif_syms
)

# S3 method for data.frame
format_p_values(
  p,
  digits_p = 3,
  cols = NULL,
  ...,
  ss = signif_syms,
  signif_stars = TRUE,
  rm_zero = FALSE,
  add_p = FALSE,
  rm_spaces = FALSE
)

add_signif_stars(p, ss = signif_syms)

get_signif_stars(p, ss = signif_syms)

signif_syms_05s

p05

signif_syms_01s

p01

signif_syms_001s

p001

signif_syms_001

signif_syms_01

signif_syms_05

signif_syms

p05_01_001

p05plus

signif_stars_legend_2(ss = signif_syms)

signif_stars_legend(
  ss = signif_syms,
  decreasing = FALSE,
  collapse = c("  \n", ", ", "; ")
)

rm_zero(str, dec = ".")

format_as_p_columns(
  data,
  colnames = c("p.value", "p.adjust"),
  digits_p = 3,
  rm_zero = FALSE,
  signif_stars = FALSE,
  ...
)

Format

An object of class numeric of length 1.

An object of class numeric of length 1.

An object of class numeric of length 1.

An object of class numeric of length 1.

An object of class numeric of length 1.

An object of class numeric of length 1.

An object of class numeric of length 1.

An object of class numeric of length 2.

An object of class numeric of length 3.

An object of class numeric of length 4.

An object of class numeric of length 4.

An object of class numeric of length 4.

Arguments

p

Un (vector de) valores de p. Numerico. O un data.frame.

digits_p

(numerico) Número de dígitos significativos para redondear un valor p. No menos de 2.

if digits_p = 2:

los valores inferiores a 0.001 se imprimen como "<0.001";

valores entre 0.001 y 0.01 como "<0.01";

todos los demás valores se redondean a dos decimales.

if digits_p = 3, sólo formato "<0.01" se salta.

cols, alpha, ss, decreasing, collapse, dec

# FIXME: not documented

...

Argumentos para otros métodos.

signif_stars

(logical) Marca si se deben añadir estrellas de significación a cada valor p. No menos de 2.

rm_zero

(logical) Marca si el cero inicial antes del punto debe ser removido.

add_p

(logical) Marca si la letra "p" debe incluirse en la expresión.

rm_spaces

(logical) Marca si se deben eliminar todos los espacios.

Value

Un vector de caracteres con valores p formateados.

Details

format_p - formatos único valor p.

format_as_p_columns - formatea las columnas numericas indicadas en un marco de datos como valores p (las columnas se convierten en cadenas).

get_signif_stars - toma los valores numéricos de p aporta las estrellas de significación estadística apropiadas.

add_signif_stars - formatea los valores p numéricos añadiendo estrellas de significación (el resultado es un vector de caracteres).

signif_stars_legend - genera una leyenda para las estrellas de significación (el resultado es una cadena).

rm_zero - elimina el cero al principio de un número (devuelve una cadena con el mismo valor pero sin el cero inicial).

Note

Función tomada y traducida de la paqueteria Biostat: https://github.com/GegznaV/biostat.

Examples

# Prettify p-values

#format_p_values(0.0005)

#format_p_values(0.005)
#format_p_values(0.005, signif_stars = FALSE)
#format_p_values(0.005, rm_zero = TRUE)
#format_p_values(0.005, digits_p = 2)
#format_p_values(0.005, digits_p = 2, rm_zero = TRUE, signif_stars = FALSE)

#format_p_values("0.00022")
#format_p_values("0.052")

#format_p_values(c(0.005, 0.0005, 0.052147))

#get_signif_stars(0.005)
#add_signif_stars(0.005)

#get_signif_stars(0.0005)
#add_signif_stars(0.0005)

#get_signif_stars(0.052147)
#add_signif_stars(0.052147)

#signif_stars_legend()

#format_p(0)
#format_p(.02, digits_p = 2)
#format_p(.0002)
#format_p(.0002, signif_stars = FALSE)
#format_p(.0002, ss = c("*****" = 0.001))

# TODO:
# 1. Añadir parámetro para activar la corrección del valor p
#    de p = 1 a, por ejemplo, p > 0,999;
#
# 2. fusionar parámetros `ss` y `signif_stars`
# 3. test: format_p(NaN) -- [OK]
# 4. test: format_p(NA) -- esta función falla con NA como entrada.
rm_zero(0.020)
#> [1] ".02"