Limpia un vector de texto, que suele contener los nombres de un objeto.

Los vectores resultantes son unicos y estan formadas unicamente por el caracter _, numeros y letras. Por defecto, solo consistiran en caracteres ASCII, pero se puede permitir que no sean ASCII (por ejemplo, Unicode) configurando ascii=FALSE. Las preferencias de mayusculas pueden especificarse utilizando el parametro case.

Cuando ascii=TRUE (el valor predeterminado), los caracteres acentuados se transliteran a ASCII. Por ejemplo, una "o" con dieresis alemana se convierte en "o", y el caracter español "enye" se convierte en "n". Esta funcion fue tomada del paquete janitor.

Usage

limpiar_nombres2(
  string,
  case = "snake",
  replace = c(`'` = "", `"` = "", `%` = "_percent_", `#` = "_number_"),
  ascii = TRUE,
  use_make_names = TRUE,
  sep_in = "\\.",
  transliterations = "Latin-ASCII",
  parsing_option = 1,
  numerals = "asis",
  ...
)

Arguments

string

Un vector de caracteres de nombres para limpiar.

case

Preferencias de mayusculas

replace

Un vector de caracteres con nombre en el que el nombre se sustituye por el value.

ascii

Convertir los nombres a ASCII (TRUE, por defecto) o no (FALSE).

use_make_names

¿Deberia aplicarse el codigo make.names() para asegurar que la sea utilizable como un nombre sin comillas? (Evitar make.names() asegura que la salida es independiente de la localizacion, pero las comillas pueden ser necesarias).

sep_in

(abreviatura de entrada separadora) si es un carácter, se interpreta como una expresión regular (envuelta internamente en stringr::regex()). El valor por defecto es una expresión regular que coincide con cualquier secuencia de valores no alfanuméricos. Todas las coincidencias serán reemplazadas por guiones bajos (además de "_" y " ", para los que esto siempre es cierto, incluso si se proporciona NULL). Estos guiones bajos se utilizan internamente para dividir las cadenas en subcadenas y especificar los límites de las palabras.

transliterations

Un vector de caracteres (si no es NULL). Las entradas de este argumento deben ser elementos de stringi::stri_trans_list() (como "Latin-ASCII", que suele ser útil) o nombres de tablas de búsqueda (actualmente sólo se admite "german").

parsing_option

Un entero que determinará la parsing_option.

numerals

Carácter que especifica la alineación de los numerales ("medio", izquierda, derecha, asis o apretado). Es decir, numerales = "izquierda" garantiza que no haya ningún separador de salida delante de un dígito.

...

...

Value

Devuelve el vector de caracteres "limpio".

See also

Examples

# limpiar los nombres de un vector:
x <- structure(1:3, names = c("nombre con espacio", "DosPalabras", "total $ (2009)"))
x
#> nombre con espacio        DosPalabras     total $ (2009) 
#>                  1                  2                  3 
names(x) <- limpiar_nombres2(names(x))
x # Ya tiene los nombres limpios
#> nombre_con_espacio       dos_palabras         total_2009 
#>                  1                  2                  3